Add stylesheets, revise serialization
inject-whitespace.xsl is for working with grammars translated from
other forms in which a lexer is assumed to have eaten all whitespace.
It was written for the work with the Oberon grammar.

ixml-inject-regex adds a @regex attribute to character-set terminals.
It was written to help simplify work on test-set coverage.

make-all-states-final takes a regular grammar (e.g. a regular
approximation of a CFG) and makes all states accept states.  That
means it can be used to generate coverage reports for both positive
and negative test cases.

serialize-as-ixml has been updated to use current ixml element names.